move fieldrefs to their own .cpp file
